Fixed AddWatchList so that the constraint failures are simply treated as user aborts

+++ src/benchmarks/edu/brown/benchmark/wikipedia/util/WikipediaUtil.java
@@ -79,7 +79,7 @@ public final class WikipediaUtil {
      * @param pageId
      * @return
      */
-    public int getPageNameSpace(int pageId) {
+    public int getPageNameSpace(long pageId) {
         return (0);
     }
